// src/types/search.d.ts
// Generated by https://quicktype.io
//
// To change quicktype's target language, run command:
//
//   "Set quicktype target language"

export interface searchResult {
  page: number;
  results: Result[];
  total_pages: number;
  total_results: number;
}

export interface Result {
  adult: boolean;
  id: number;
  name?: string;
  original_name?: string;
  media_type: string;
  popularity: number;
  gender?: number;
  known_for_department?: string;
  profile_path?: string;
  known_for?: KnownFor[];
  backdrop_path?: null | string;
  title?: string;
  original_language?: string;
  original_title?: string;
  overview?: string;
  poster_path?: null | string;
  genre_ids?: number[];
  release_date?: string;
  video?: boolean;
  vote_average?: number;
  vote_count?: number;
}

export interface KnownFor {
  adult: boolean;
  backdrop_path: string;
  id: number;
  title: string;
  original_language: string;
  original_title: string;
  overview: string;
  poster_path: string;
  media_type: string;
  genre_ids: number[];
  popularity: number;
  release_date: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
}